Refactor: Move hero and item prices to initState
This commit moves the initialization of `heroPrices` and `itemPrice` maps from the class level to the `initState` method in `_ShopScreenState`.

Additionally, added TODO comments to indicate that these prices should eventually be loaded from an external file.
Minor formatting changes were also applied to the `ItemCard` widgets for better readability.